requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
gitdb==4.0.11
GitPython==3.1.43
idna==3.8
# Editable install with no version control (LLMAnalyzer==0.1)
-e /Users/yavital/dev/llm-analyzer
requests==2.32.3
setuptools==75.1.0
smmap==5.0.1
urllib3==2.2.2
